+++ src/main/java/com/expense/model/SpendSummary.java
@@ -1,33 +1,37 @@
 package com.expense.model;
-
+/**
+ * Spend Summary
+ * @author USER
+ *
+ */
 public class SpendSummary {
 	
 	private String groupId;
-	private int totalCost;
-	private int perPsersonCost;
-	private int pendingSettlement;
+	private double totalCost;
+	private double perPsersonCost;
+	private double pendingSettlement;
 	public String getGroupId() {
 		return groupId;
 	}
-	public int getTotalCost() {
+	public double getTotalCost() {
 		return totalCost;
 	}
-	public int getPerPsersonCost() {
+	public double getPerPsersonCost() {
 		return perPsersonCost;
 	}
-	public int getPendingSettlement() {
+	public double getPendingSettlement() {
 		return pendingSettlement;
 	}
 	public void setGroupId(String groupId) {
 		this.groupId = groupId;
 	}
-	public void setTotalCost(int totalCost) {
+	public void setTotalCost(double totalCost) {
 		this.totalCost = totalCost;
 	}
-	public void setPerPsersonCost(int perPsersonCost) {
+	public void setPerPsersonCost(double perPsersonCost) {
 		this.perPsersonCost = perPsersonCost;
 	}
-	public void setPendingSettlement(int pendingSettlement) {
+	public void setPendingSettlement(double pendingSettlement) {
 		this.pendingSettlement = pendingSettlement;
 	}
